+++ tests/testthat/test_variable_importance.R
@@ -1,8 +1,8 @@
 context("Check variable_importance() function")
 
-vi_ranger <- variable_importance(explainer_regr_ranger, n_sample = 100)
-vi_lm <- variable_importance(explainer_regr_lm, n_sample = 100)
-vi_ranger_ratio <- variable_importance(explainer_regr_ranger, n_sample = 100, type = "ratio")
+vi_ranger <- variable_importance(explainer_regr_ranger, N = 100)
+vi_lm <- variable_importance(explainer_regr_lm, N = 100)
+vi_ranger_ratio <- variable_importance(explainer_regr_ranger, N = 100, type = "ratio")
 
 test_that("Description prints properly", {
   des <- ingredients::describe(vi_ranger)
@@ -10,11 +10,11 @@ test_that("Description prints properly", {
 })
 
 test_that("y not provided",{
-  expect_error(variable_importance(explainer_regr_ranger_wo_y, n_sample = 100))
+  expect_error(variable_importance(explainer_regr_ranger_wo_y, N = 100))
 })
 
 test_that("data not provided",{
-  expect_error(variable_importance(explainer_wo_data, n_sample = 100))
+  expect_error(variable_importance(explainer_wo_data, N = 100))
 })
 
 test_that("wrong type value",{
@@ -37,5 +37,3 @@ test_that("Inverse sorting of bars",{
   expect_is(plot(vi_ranger, vi_lm, desc_sorting = FALSE), "gg")
   expect_is(plot(vi_ranger, desc_sorting = FALSE), "gg")
 })
-
-